+++ CHANGELOG.rst
@@ -8,6 +8,7 @@ Changelog
 
 - Remove Document to Collection link
 - Rename Unicode field to String
+- Extract AbstractField
 
 0.3.0 (2015-04-27)
 ++++++++++++++++++